+++ app/src/main/java/com/wb/logistics/network/client/OkHttpClientUnsafe.kt
@@ -21,8 +21,8 @@ object OkHttpClientUnsafe {
             )
             .hostnameVerifier { _: String?, _: SSLSession? -> true }
             .addInterceptor(httpLoggerInterceptor)
-            .connectTimeout(AppConfig.HTTP_CONNECT_TIMEOUT.toLong(), TimeUnit.MILLISECONDS)
-            .readTimeout(AppConfig.HTTP_READ_TIMEOUT.toLong(), TimeUnit.MILLISECONDS)
+            .connectTimeout(AppConfig.HTTP_CONNECT_TIMEOUT, TimeUnit.MILLISECONDS)
+            .readTimeout(AppConfig.HTTP_READ_TIMEOUT, TimeUnit.MILLISECONDS)
         return okHttpBuilder.build()
     }
 
@@ -40,8 +40,8 @@ object OkHttpClientUnsafe {
             .hostnameVerifier { _: String?, _: SSLSession? -> true }
             .addInterceptor(refreshResponseInterceptor)
             .addInterceptor(httpLoggerInterceptor)
-            .connectTimeout(AppConfig.HTTP_CONNECT_TIMEOUT.toLong(), TimeUnit.MILLISECONDS)
-            .readTimeout(AppConfig.HTTP_READ_TIMEOUT.toLong(), TimeUnit.MILLISECONDS)
+            .connectTimeout(AppConfig.HTTP_CONNECT_TIMEOUT, TimeUnit.MILLISECONDS)
+            .readTimeout(AppConfig.HTTP_READ_TIMEOUT, TimeUnit.MILLISECONDS)
         return okHttpBuilder.build()
     }
